REF: move coverage_simplify/coverage_is_valid implementation
