+++ docs/coverage.rst
@@ -1,0 +1,11 @@
+Constructive operations
+=======================
+
+.. currentmodule:: shapely
+
+.. autosummary::
+   :toctree: reference/
+
+{% for function in get_module_functions("_coverage") %}
+   {{ function }}
+{% endfor %}
